+++ config/config.json
@@ -1,7 +1,7 @@
 {
   "development": {
     "username": "root",
-    "password": null,
+    "password": "root",
     "database": "database_development",
     "host": "127.0.0.1",
     "dialect": "mysql"
