+++ routes/api-routes.js
@@ -4,23 +4,30 @@ var db = require("../models");
 // =============================================================
 module.exports = function(app) {
 
-  app.get("*", function(req, res) {
-    res.render("index", {data:"Hello World!"})
-  });
+ 
 
 
   app.get("/admin/:table/:category?", function(req,res){
+    console.log(req.params);
     let table=req.params.table;
-    if(req.params.category){
-      let category=req.params.category;
-      db.table.findAll({
-        attributes: category,
-        })
-    }else{
-      db.table.findAll().then(function(results){
+    // if(req.params.category){
+    //   let category=req.params.category;
+    //   db.todo.findAll({
+    //     attributes: category,
+    //     }).then(function(results){
+    //       res.json(results);
+    //     });
+    // }else{
+      db.todo.findAll({}).then(function(results){
         res.json(results);
-      })
-    }
-  }
+      });
+    // }
+  });
 
-};
+
+
+  app.get("*", function(req, res) {
+    res.render("index", {data:"Hello World!"})
+  });
+}
+
