--- sever.js
@@ -1,34 +1,0 @@
-// Server.js - This file is the initial starting point for the Node/Express server.
-
-// Dependencies
-// =============================================================
-var express = require("express");
-
-// Sets up the Express App
-// =============================================================
-var app = express();
-var PORT = process.env.PORT || 8080;
-
-// Sets up the Express app to handle data parsing
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-
-// Static directory to be served
-app.use(express.static("app/public"));
-// Set Handlebars.
-var exphbs = require("express-handlebars");
-
-app.engine("handlebars", exphbs({ defaultLayout: "main" }));
-app.set("view engine", "handlebars");
-// Routes
-// =============================================================
-require("./routes/api-routes.js")(app);
-
-// Here we introduce HTML routing to serve different HTML files
-// require("./routes/html-routes.js")(app);
-
-// Starts the server to begin listening
-// =============================================================
-app.listen(PORT, function() {
-  console.log("App listening on PORT " + PORT);
-});
